Changed which tool I use to convert .bin to .8xp
The old BinTo8xpEng tool was unacceptable because it only had a GUI
interface.
I wanted to use Bin2Var, but it wouldn't let me change the on-calculator
program name.
The tool I'm using now still has issues: it requires Python, and it also
has some bugs in the way it deals with files in different folders, which
is why it is placed in the bin/ folder for now.